Silver1 dev, working version for predifined devices

+++ Autotuning_runtime/include/CoCoPeLiaModelLvl1.hpp
@@ -9,7 +9,7 @@
 #include "CoCoPeLiaCoModel.hpp"
 #include "backend_wrappers.hpp"
 
-CoCoModel_p CoCoModelFuncInitBLAS1(CoCoModel_p out_model, short dev_id, const char* func, void* func_data);
+CoCoModel_p CoCoModelFuncInitBLAS1(CoCoModel_p out_model, int dev_id, const char* func, void* func_data);
 
 long int CoCopeLiaMinAllowedTBLAS1(CoCoModel_p model);
 long int CoCopeLiaMaxAllowedTBLAS1(CoCoModel_p model);
@@ -25,7 +25,7 @@ double CoCopeLiaPredictDataLocBLAS1(CoCoModel_p model, long int T);
 double CoCopeLiaPredictBidirectionalBLAS1(CoCoModel_p model, long int T);
 double CoCopeLiaPredictReuseBLAS1(CoCoModel_p model, long int T);
 double CoCopeLiaPipelineEmulateBLAS1(CoCoModel_p model, long int T);
-double CoCopeLiaPredictBidirectionalHeteroBLAS1(CoCoModel_p model, short used_devs, short* used_dev_ids,
+double CoCopeLiaPredictBidirectionalHeteroBLAS1(CoCoModel_p model, int used_devs, int* used_dev_ids,
 	double* used_dev_relative_scores, long int T);
-  
+
 #endif
